// src/app/types/artist-waitlist-form.d.ts
export interface IArtistWaitlistForm {
  from_name: string;
  from_email: string;
  social_handle: string;
  artistic_background: string;
  creative_work_link: string;
  collections_link: string;
  technical_proficiency: string;
  project_concept: string;
  social_link: string;
  followers_count: string;
  project_stage: string;
  i_certify: boolean;
}